rename 'boundingbox' to 'core'
