+++ barcode_detection/core/bounding_box.py
