+++ barcode_detection/decoding/zxing_method/decode_zxing.py
@@ -3,7 +3,7 @@ import docker
 import numpy as np
 import tempfile
 
-from barcode_detection.boundingbox.bounding_box import BoundingBox
+from barcode_detection.core.bounding_box import BoundingBox
 from barcode_detection.decoding.decode import Decoder
 from barcode_detection.utils import crop_img
 from pathlib import Path
